+++ app/tools/loggers.py
@@ -3,23 +3,88 @@
 import os
 import logging
 from logging import StreamHandler
-from logging.handlers import SMTPHandler, RotatingFileHandler
+from logging.handlers import TimedRotatingFileHandler
 
 import flask
 from discord_webhook import DiscordWebhook
 
 
 class DiscordHandler(StreamHandler):
-    def __init__(self, webhook, role_id=None):
+    """Logging handler using a webhook to send log to a Discord server
+
+    Args:
+        webhook (str): Webhook ID to use
+            ("https://discord.com/api/webhooks/<server>/<id>")
+    """
+    def __init__(self, webhook):
+        """Initializes self."""
         super().__init__()
         self.webhook = webhook
-        self.gri_mention = f"<@&{role_id}> " if role_id else ""
 
     def emit(self, record):
-        msg = self.format(record)
+        """Method called to make this handler send a record."""
+        content = self.format(record)
+        webhook = DiscordWebhook(url=self.webhook, content=content)
+        webhook.execute()
+
+
+class InfoErrorFormatter(logging.Formatter):
+    """Utility formatter allowing to use different formats for info and errors.
+
+    Args:
+        info_fmt (str): Formatter format to use for levels DEBUG and INFO.
+        error_fmt (str): Formatter format to use for levels above INFO.
+        *args, **kwargs: Passed to :class:`logging.Formatter`
+    """
+    def __init__(self, info_fmt, error_fmt, *args, **kwargs):
+        """Initializes self."""
+        super().__init__(info_fmt, *args, **kwargs)
+        self.info_fmt = info_fmt
+        self.error_fmt = error_fmt
+
+    def format(self, record):
+        """Method called to make this formatter format a record."""
+        format_orig = self._fmt
+        if record.levelno > logging.INFO:
+            self._fmt = self.error_fmt
+        else:
+            self._fmt = self.info_fmt
+        result = super().format(record)
+        self._fmt = format_orig
+        return record.name.ljust(12) + result
+
+
+class DiscordFormatter(logging.Formatter):
+    """Base formatter preparing a record to be send to a Discord server.
+
+    Args:
+        role_id (str): Optional 18-digit ID used to mention a specific role
+            in subclasses using :attr:`DiscordFormatter.role_mention`.
+    """
+    def __init__(self, role_id=None):
+        """Initializes self."""
+        super().__init__()
+        self.role_mention = f"<@&{role_id}> " if role_id else ""
+
+    def format(self, record):
+        """Method called to make this formatter format a record.
+
+        Truncates message to 1900 characters (Discord limits to 2000).
+        """
+        msg = record.getMessage()
         if len(msg) > 1900:     # Discord limitation
             msg = "[...]\n" + msg[-1900:]
+        return msg
 
+
+class DiscordErrorFormatter(DiscordFormatter):
+    """:class:`.DiscordFormatter` subclass used to transmit error messages."""
+    def format(self, record):
+        """Method called to make this formatter format a record.
+
+        Retrieves request IP and logged-in rezident name if applicable.
+        """
+        msg = super().format(record)
         try:
             remote_ip = flask.g.remote_ip or "<missing header>"
         except AttributeError:
@@ -31,61 +96,64 @@ class DiscordHandler(StreamHandler):
             except AttributeError:
                 pass
 
-        content = (f"{self.gri_mention}ALED ça a planté ! (chez {remote_ip})\n"
-                   f"```{msg}```")
-        webhook = DiscordWebhook(url=self.webhook, content=content)
-        webhook.execute()
-
+        return (f"{self.role_mention}ALED ça a planté ! (chez {remote_ip})\n"
+                f"```{msg}```")
 
 # Custom formatter
-class InfoErrorFormatter(logging.Formatter):
-    def __init__(self, info_fmt, error_fmt, *args, **kwargs):
-        super().__init__(info_fmt, *args, **kwargs)
-        self.info_fmt = info_fmt
-        self.error_fmt = error_fmt
-
+class DiscordLoggingFormatter(DiscordFormatter):
+    """:class:`.DiscordFormatter` subclass used to transmit actions infos."""
     def format(self, record):
-        format_orig = self._fmt
-        if record.levelno > logging.INFO:
-            self._fmt = self.error_fmt
+        """Method called to make this formatter format a record.
+
+        Retrieves logged-in rezident and doaser names if applicable.
+        """
+        msg = super().format(record)
+        if flask.g.logged_in:
+            user = flask.g.logged_in_user.username
+            if flask.g.doas:
+                user += f" AS {flask.g.rezident.username}"
         else:
-            self._fmt = self.info_fmt
-        result = super().format(record)
-        self._fmt = format_orig
-        return result
+            user = "(anonymous)"
+        if record.levelno > logging.INFO:
+            return f"`{user}: {record.levelname}: {msg}` ({self.role_mention})"
+        else:
+            return f"`{user}: {msg}`"
 
 
-def set_handlers(app):
-    if app.config["MAIL_SERVER"] and not (app.debug or app.testing):
-        # Alert mails
-        server = app.config["MAIL_SERVER"]
-        auth = None
-        if app.config["MAIL_USERNAME"] or app.config["MAIL_PASSWORD"]:
-            auth = (app.config["MAIL_USERNAME"], app.config["MAIL_PASSWORD"])
-        secure = () if app.config["MAIL_USE_TLS"] else None
-        mail_handler = SMTPHandler(
-            mailhost=(server, app.config["MAIL_PORT"]),
-            fromaddr=f"no-reply@{server}",
-            toaddrs=app.config["ADMINS"],
-            subject="IntraRez Internal Error",
-            credentials=auth,
-            secure=secure
-        )
-        mail_handler.setLevel(logging.ERROR)
-        app.logger.addHandler(mail_handler)
+def configure_logging(app):
+    """Configure logging for the IntraRez web app.
+
+    Setup :attr:`app.logger <flask.Flask.logger>` to log errors to
+    ``app.config["ERROR_WEBHOOK"]`` Discord webhook and everything to a
+    journalized file, and adds a child logger  ``app.actions_logger``
+    ("app.actions") logging actions to ``app.config["LOGGING_WEBHOOK"]``.
+    """
+    # Add rezidents actions logger
+    app.actions_logger = app.logger.getChild("actions")
 
     if app.config["ERROR_WEBHOOK"] and not (app.debug or app.testing):
-        # Alert webhooks
-        discord_handler = DiscordHandler(app.config["ERROR_WEBHOOK"],
-                                         app.config.get("GRI_ROLE_ID"))
-        discord_handler.setLevel(logging.ERROR)
-        app.logger.addHandler(discord_handler)
+        # Alert messages for errors
+        discord_errors_handler = DiscordHandler(app.config["ERROR_WEBHOOK"])
+        discord_errors_handler.setLevel(logging.ERROR)
+        discord_errors_handler.setFormatter(DiscordErrorFormatter(
+            app.config.get("GRI_ROLE_ID")
+        ))
+        app.logger.addHandler(discord_errors_handler)
 
-    # File logs
+    if app.config["LOGGING_WEBHOOK"]:
+        # Logging messages for actions
+        discord_actions_handler = DiscordHandler(app.config["LOGGING_WEBHOOK"])
+        discord_actions_handler.setLevel(logging.DEBUG if app.debug
+                                         else logging.INFO)
+        discord_actions_handler.setFormatter(DiscordLoggingFormatter(
+            app.config.get("GRI_ROLE_ID")
+        ))
+        app.actions_logger.addHandler(discord_actions_handler)
+
+    # File logging
     if not os.path.exists("logs"):
         os.mkdir("logs")
-    file_handler = RotatingFileHandler("logs/intrarez.log", maxBytes=10240,
-                                       backupCount=100)
+    file_handler = TimedRotatingFileHandler("logs/intrarez.log", when="D")
     file_handler.setFormatter(InfoErrorFormatter(
         "{asctime} {levelname}: {message}",
         "{asctime} {levelname}: {message} [in {pathname}:{lineno}]",
@@ -96,3 +164,4 @@ def set_handlers(app):
 
     # Start logs
     app.logger.setLevel(logging.INFO)
+    app.actions_logger.setLevel(logging.INFO)